+++ compose/modules/localnet/docker/splice/health-check.sh
@@ -15,3 +15,6 @@ if [ "$SV_VALIDATOR_1_PROFILE" = "on" ]; then
   curl -f http://localhost:5012/api/scan/readyz
   curl -f http://localhost:5014/api/sv/readyz
 fi
+if [ "$D_VALIDATOR_1_PROFILE" = "on" ]; then
+  curl -f "http://localhost:14${VALIDATOR_ADMIN_API_PORT_SUFFIX}/api/validator/readyz"
+fi