+++ compose/modules/splice-onboarding/docker/health-check.sh
@@ -44,6 +44,14 @@ if [ ! -f /tmp/all-done ]; then
     fi
   fi
 
+  if [ "$D_VALIDATOR_1_PROFILE" == "on" ]; then
+    source /app/d-validator-1-auth.sh
+    if [ "$DO_INIT" == "true" ] && [ ! -f /tmp/d-validator-1-init-dars-uploaded ]; then
+      upload_dars "$D_VALIDATOR_1_PARTICIPANT_ADMIN_TOKEN" "canton:14${PARTICIPANT_JSON_API_PORT_SUFFIX}"
+      touch /tmp/d-validator-1-init-dars-uploaded
+    fi
+  fi
+
   echo "Executing onboarding scripts..." >&2
 
   for script in $(ls "$ONBOARDING_SCRIPTS_DIR"); do